move shared helpers to top shared module. fix fetch_data return in get_artifacts with new helper
